fix: remove title from Card and move it to TokenCard
