+++ src/components/Token/TokenList/TokenCard.tsx
@@ -5,22 +5,22 @@ import RoutePaths from '@src/shared/enums/routes-paths'
 import { IpfsTokenContent } from '@src/shared/types/ipfs'
 import { getTextContentFromHtml, limitString } from '@src/shared/utils'
 import { generatePath } from 'react-router-dom'
-import { usePageCommentsCount } from '@src/hooks/subgraph/usePageCommentsCount'
-import styled from 'styled-components'
+import styled, { useTheme } from 'styled-components'
 import Flex from '@src/components/ui/Flex'
+import { useTokenCommentsCount } from '@src/hooks/subgraph/useTokenCommentsCount'
+import RouterLink from '@src/components/ui/RouterLink'
 
 const CountWrapper = styled(Flex)`
-  position: absolute;
-  top: 12px;
-  right: 16px;
   background-color: ${({ theme }) => theme.palette.white};
   color: ${({ theme }) => theme.palette.linkPrimary};
   padding: 8px 12px;
   z-index: 1;
 `
 
-const CardWrapper = styled(Flex)`
-  position: relative;
+const Title = styled(Text.h2)`
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
 `
 
 interface TokenCardProps {
@@ -36,27 +36,40 @@ const TokenCard: React.FC<TokenCardProps> = ({
   content,
   name,
 }) => {
-  const commentsCount = usePageCommentsCount(tokenId)
+  const commentsCount = useTokenCommentsCount(tokenId)
+  const theme = useTheme()
 
   return (
-    <CardWrapper>
-      <Card
-        to={generatePath(RoutePaths.NFT + RoutePaths.TOKEN, {
-          nftId,
-          tokenId: tokenId,
-        })}
-        title={name}
+    <Card>
+      <Flex
+        justifyContent='space-between'
+        alignItems='center'
+        marginBottom='5px'
       >
-        {content && (
-          <Text.p>
-            {limitString(getTextContentFromHtml(content.htmlContent), 700)}
-          </Text.p>
-        )}
-      </Card>
-      <CountWrapper>
-        <Text.p color='color'>Attestations: {commentsCount}</Text.p>
-      </CountWrapper>
-    </CardWrapper>
+        <RouterLink
+          to={generatePath(RoutePaths.NFT + RoutePaths.TOKEN, {
+            nftId,
+            tokenId: tokenId,
+          })}
+        >
+          <Title
+            color={theme.palette.linkPrimary}
+            size={theme.fontSizes.mediumPlus}
+            weight={theme.fontWeights.bold}
+          >
+            {name}
+          </Title>
+        </RouterLink>
+        <CountWrapper>
+          <Text.p color='color'>Attestations: {commentsCount}</Text.p>
+        </CountWrapper>
+      </Flex>
+      {content && (
+        <Text.p>
+          {limitString(getTextContentFromHtml(content.htmlContent), 700)}
+        </Text.p>
+      )}
+    </Card>
   )
 }
